Fixed some bugs in script

+++ scripts/rebuild_halo_table_cache_log.py
@@ -2,10 +2,10 @@
 
 """Command-line script to rebuild the halo table cache log"""
 
-import argparse, os
+import argparse, os, fnmatch
 from astropy.table import Table
 
-from .halotools.sim_manager import manipulate_cache_log
+from halotools.sim_manager import manipulate_cache_log
 
 
 fname_cache_log = manipulate_cache_log.get_halo_table_cache_log_fname()
@@ -23,14 +23,16 @@ def fnames_in_existing_log():
 def halo_table_fnames_in_standard_cache():
     """
     """
-    standard_loc = os.path.join(os.path.dirname(fname_cache_log), 'halo_tables')
+    standard_loc = os.path.join(os.path.dirname(fname_cache_log), 'halo_catalogs')
     if os.path.exists(standard_loc):
-        for path, dirlist, filelist in os.walk(dirname_halo_tables):
+        for path, dirlist, filelist in os.walk(standard_loc):
             for name in fnmatch.filter(filelist, '*.hdf5'):
                 yield os.path.join(path, name)
 
+print("Number of files in standard cache loc = " + str(len(list(halo_table_fnames_in_standard_cache()))))
+
 potential_fnames = fnames_in_existing_log()
-potential_fnames.extend(list(halo_table_fnames_in_standard_cache))
+potential_fnames.extend(list(halo_table_fnames_in_standard_cache()))
 potential_fnames = list(set(potential_fnames))
 
 def verified_fname_generator():
@@ -45,10 +47,8 @@ def verified_fname_generator():
 
 verified_fnames = list(verified_fname_generator())
 
-try:
+if os.path.exists(fname_cache_log):
     os.system('rm ' + fname_cache_log)
-except OSError:
-    pass
 
 new_log = Table()
 new_log['fname'] = verified_fnames
